refactor: split up test code for clipboard lib
