+++ crates/clipboard-lib/src/lib.rs
@@ -1,1 +1,1 @@
-pub mod test_clipboard;
+pub mod store;
